Load user history on page load
Man this needs to be re-written. So added a call to history to get completed exercise, but really this should be loaded parallel to the user and the front-end is a mess in terms of parallelizing and cacheability. May slow things down horrendously in prod due to more history data and lack of converting from Guest to real user